--- src/0xsquid/index.ts
@@ -1,5 +1,0 @@
-export * from "./v1/status";
-export * from "./v2/sdk-info";
-export * from "./v2/route";
-export * from "./v2/chains";
-export * from "./v2/tokens";
