--- src/0xsquid/v2/chains.spec.ts
@@ -1,360 +1,0 @@
-import {
-  ChainName,
-  ChainType,
-  CosmosChain,
-  NetworkIdentifier
-} from "@0xsquid/squid-types";
-import {
-  parseEvmChain,
-  parseBaseChain,
-  parseChainData,
-  parseCosmosChain,
-  parseSquidContracts,
-  parseAxelarContracts
-} from "./chains";
-
-describe("chains", () => {
-  const data = {
-    chains: [
-      {
-        axelarChainName: ChainName.ETHEREUM,
-        networkIdentifier: NetworkIdentifier.ETHEREUM,
-        chainType: ChainType.EVM,
-        rpc: "https://eth-rpc.gateway.pokt.network",
-        networkName: "Ethereum",
-        chainId: "1",
-        nativeCurrency: {
-          name: "Ethereum",
-          symbol: "ETH",
-          decimals: 18,
-          icon: "https://tokens.1inch.io/0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee.png"
-        },
-        swapAmountForGas: "2000000",
-        chainIconURI:
-          "https://tokens.1inch.io/0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee.png",
-        blockExplorerUrls: ["https://etherscan.io/"],
-        chainNativeContracts: {
-          wrappedNativeToken: "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
-          ensRegistry: "0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e",
-          multicall: "0x5e227AD1969Ea493B43F840cfF78d08a6fc17796",
-          usdcToken: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48"
-        },
-        axelarContracts: {
-          gateway: "0x4F4495243837681061C4743b74B3eEdf548D56A5",
-          forecallable: ""
-        },
-        squidContracts: {
-          squidRouter: "0xce16F69375520ab01377ce7B88f5BA8C48F8D666",
-          defaultCrosschainToken: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
-          squidMulticall: "0x4fd39C9E151e50580779bd04B1f7eCc310079fd3"
-        },
-        estimatedRouteDuration: 960,
-        estimatedExpressRouteDuration: 20
-      },
-      {
-        axelarChainName: ChainName.EVMOS,
-        networkIdentifier: NetworkIdentifier.EVMOS,
-        chainType: ChainType.COSMOS,
-        rpc: "/chain/evmos",
-        rest: "https://rest.bd.evmos.org:1317",
-        nativeCurrency: {
-          name: "Evmos",
-          symbol: "EVMOS",
-          decimals: 6,
-          icon: "https://raw.githubusercontent.com/cosmos/chain-registry/master/evmos/images/evmos.png"
-        },
-        swapAmountForGas: "2000000",
-        chainIconURI: "https://axelarscan.io/logos/chains/evmos.svg",
-        networkName: "Evmos",
-        chainId: "evmos_9001-2",
-        blockExplorerUrls: [
-          "https://ezstaking.tools/evmos",
-          "https://www.mintscan.io/evmos",
-          "https://evmos.bigdipper.live",
-          "https://evm.evmos.org",
-          "https://ping.pub/evmos",
-          "https://evmos.explorers.guru",
-          "https://atomscan.com/evmos",
-          "https://evmos.tcnetwork.io"
-        ],
-        stakeCurrency: {
-          coinDenom: "EVMOS",
-          coinMinimalDenom: "aevmos",
-          coinDecimals: 18,
-          coingeckoId: "evmos"
-        },
-        walletUrl: "https://wallet.keplr.app/chains/evmos",
-        walletUrlForStaking: "https://wallet.keplr.app/chains/evmos",
-        bip44: {
-          coinType: 60
-        },
-        bech32Config: {
-          bech32PrefixAccAddr: "evmos",
-          bech32PrefixAccPub: "evmospub",
-          bech32PrefixValAddr: "evmosvaloper",
-          bech32PrefixValPub: "evmosvaloperpub",
-          bech32PrefixConsAddr: "evmosvalcons",
-          bech32PrefixConsPub: "evmosvalconspub"
-        },
-        currencies: [
-          {
-            coinDenom: "EVMOS",
-            coinMinimalDenom: "aevmos",
-            coinDecimals: 18,
-            coingeckoId: "evmos"
-          }
-        ],
-        feeCurrencies: [
-          {
-            coinDenom: "EVMOS",
-            coinMinimalDenom: "aevmos",
-            coinDecimals: 18,
-            coingeckoId: "evmos"
-          }
-        ],
-        gasPriceStep: {
-          low: 25000000000,
-          average: 25000000000,
-          high: 40000000000
-        },
-        coinType: 60,
-        features: ["ibc-transfer", "ibc-go", "eth-address-gen", "eth-key-sign"],
-        squidContracts: {
-          defaultCrosschainToken: "uusdc"
-        },
-        axelarContracts: {
-          gateway: ""
-        },
-        chainToAxelarChannelId: "channel-21",
-        estimatedRouteDuration: 180,
-        estimatedExpressRouteDuration: 20
-      }
-    ]
-  };
-  describe("parseSquidContracts", () => {
-    describe("exact match", () => {
-      const selected = { ...data.chains[0].squidContracts };
-      delete (selected as any).defaultCrosschainToken;
-      delete (selected as any).squidMulticall;
-      const expected = parseSquidContracts(selected);
-      it("should match provided data", () => {
-        expect(expected).toEqual(selected);
-      });
-      it("should contain squidRouter", () => {
-        expect(expected).toHaveProperty("squidRouter");
-      });
-      it("should contain defaultCrosschainToken", () => {
-        expect(expected).not.toHaveProperty("defaultCrosschainToken");
-      });
-      it("should contain squidMulticall", () => {
-        expect(expected).not.toHaveProperty("squidMulticall");
-      });
-    });
-    describe("optional properties", () => {
-      const selected = { ...data.chains[0].squidContracts };
-      const expected = parseSquidContracts(selected);
-      it("should contain defaultCrosschainToken", () => {
-        expect(expected).toHaveProperty("defaultCrosschainToken");
-      });
-      it("should contain squidMulticall", () => {
-        expect(expected).toHaveProperty("squidMulticall");
-      });
-    });
-  });
-  describe("parseAxelarContracts", () => {
-    describe("exact match", () => {
-      const selected = { ...data.chains[0].axelarContracts };
-      delete (selected as any).forecallable;
-      const expected = parseAxelarContracts(selected);
-      it("should match provided data", () => {
-        expect(expected).toEqual(selected);
-      });
-      it("should contain gateway", () => {
-        expect(expected).toHaveProperty("gateway");
-      });
-      it("should contain forecallable", () => {
-        expect(expected).not.toHaveProperty("forecallable");
-      });
-    });
-    describe("optional properties", () => {
-      const selected = { ...data.chains[0].axelarContracts };
-      const expected = parseAxelarContracts(selected);
-      it("should contain forecallable", () => {
-        expect(expected).toHaveProperty("forecallable");
-      });
-    });
-  });
-  describe("parseBaseChain", () => {
-    describe("exact match", () => {
-      const selected = { ...data.chains[0] };
-      delete selected.chainNativeContracts;
-      const expected = parseBaseChain(selected);
-      it("should match provided data", () => {
-        expect(expected).toEqual(selected);
-      });
-      it("should have Bridge type", () => {
-        expect(expected.chainType).toBe(ChainType.EVM);
-      });
-      it("should contain axelarChainName", () => {
-        expect(expected).toHaveProperty("axelarChainName");
-      });
-      it("should contain rpc", () => {
-        expect(expected).toHaveProperty("rpc");
-      });
-      it("should contain networkName", () => {
-        expect(expected).toHaveProperty("networkIdentifier");
-      });
-      it("should contain chainId", () => {
-        expect(expected).toHaveProperty("chainId");
-      });
-      it("should contain nativeCurrency", () => {
-        expect(expected).toHaveProperty("nativeCurrency");
-      });
-      it("should contain chainIconURI", () => {
-        expect(expected).toHaveProperty("chainIconURI");
-      });
-      it("should contain blockExplorerUrls", () => {
-        expect(expected).toHaveProperty("blockExplorerUrls");
-      });
-      it("should contain axelarContracts", () => {
-        expect(expected).toHaveProperty("axelarContracts");
-      });
-      it("should contain squidContracts", () => {
-        expect(expected).toHaveProperty("squidContracts");
-      });
-      it("should contain estimatedRouteDuration", () => {
-        expect(expected).toHaveProperty("estimatedRouteDuration");
-      });
-    });
-  });
-  describe("parseEvmChain", () => {
-    describe("exact match", () => {
-      const selected = data.chains[0];
-      const expected = parseEvmChain(selected);
-      it("should match provided data", () => {
-        expect(expected).toEqual(selected);
-      });
-      it("should have Bridge type", () => {
-        expect(expected.chainType).toBe(ChainType.EVM);
-      });
-      it("should contain axelarChainName", () => {
-        expect(expected).toHaveProperty("axelarChainName");
-      });
-      it("should contain rpc", () => {
-        expect(expected).toHaveProperty("rpc");
-      });
-      it("should contain networkName", () => {
-        expect(expected).toHaveProperty("networkIdentifier");
-      });
-      it("should contain chainId", () => {
-        expect(expected).toHaveProperty("chainId");
-      });
-      it("should contain nativeCurrency", () => {
-        expect(expected).toHaveProperty("nativeCurrency");
-      });
-      it("should contain chainIconURI", () => {
-        expect(expected).toHaveProperty("chainIconURI");
-      });
-      it("should contain blockExplorerUrls", () => {
-        expect(expected).toHaveProperty("blockExplorerUrls");
-      });
-      it("should contain chainNativeContracts", () => {
-        expect(expected).toHaveProperty("chainNativeContracts");
-      });
-      it("should contain axelarContracts", () => {
-        expect(expected).toHaveProperty("axelarContracts");
-      });
-      it("should contain squidContracts", () => {
-        expect(expected).toHaveProperty("squidContracts");
-      });
-      it("should contain estimatedRouteDuration", () => {
-        expect(expected).toHaveProperty("estimatedRouteDuration");
-      });
-    });
-  });
-  describe("parseCosmosChain", () => {
-    describe("exact match", () => {
-      const selected = { ...data.chains[1] };
-      delete (selected as CosmosChain).coinType;
-      delete (selected as CosmosChain).features;
-      delete (selected as CosmosChain).gasPriceStep;
-      const expected = parseCosmosChain(selected);
-      it("should match provided data", () => {
-        expect(expected).toEqual(selected);
-      });
-      it("should have Bridge type", () => {
-        expect(expected.chainType).toBe(ChainType.COSMOS);
-      });
-      it("should contain axelarChainName", () => {
-        expect(expected).toHaveProperty("axelarChainName");
-      });
-      it("should contain rpc", () => {
-        expect(expected).toHaveProperty("rpc");
-      });
-      it("should contain networkName", () => {
-        expect(expected).toHaveProperty("networkIdentifier");
-      });
-      it("should contain chainId", () => {
-        expect(expected).toHaveProperty("chainId");
-      });
-      it("should contain nativeCurrency", () => {
-        expect(expected).toHaveProperty("nativeCurrency");
-      });
-      it("should contain chainIconURI", () => {
-        expect(expected).toHaveProperty("chainIconURI");
-      });
-      it("should contain blockExplorerUrls", () => {
-        expect(expected).toHaveProperty("blockExplorerUrls");
-      });
-      it("should contain axelarContracts", () => {
-        expect(expected).toHaveProperty("axelarContracts");
-      });
-      it("should contain squidContracts", () => {
-        expect(expected).toHaveProperty("squidContracts");
-      });
-      it("should contain estimatedRouteDuration", () => {
-        expect(expected).toHaveProperty("estimatedRouteDuration");
-      });
-    });
-    describe("optional properties", () => {
-      const selected = { ...data.chains[1] };
-      const expected = parseCosmosChain(selected);
-      it("should match provided data", () => {
-        expect(expected).toEqual(selected);
-      });
-      it("should have ChainType.Cosmos", () => {
-        expect(expected.chainType).toBe(ChainType.COSMOS);
-      });
-      it("should contain optional chainId", () => {
-        expect(expected).toHaveProperty("chainId");
-      });
-      it("should contain optional coinType", () => {
-        expect(expected).toHaveProperty("coinType");
-      });
-      it("should contain optional features", () => {
-        expect(expected).toHaveProperty("features");
-      });
-      it("should contain optional blockExplorerUrls", () => {
-        expect(expected).toHaveProperty("blockExplorerUrls");
-      });
-      it("should contain optional axelarContracts", () => {
-        expect(expected).toHaveProperty("axelarContracts");
-      });
-      it("should contain optional squidContracts", () => {
-        expect(expected).toHaveProperty("squidContracts");
-      });
-      it("should contain optional gasPriceStep", () => {
-        expect(expected).toHaveProperty("gasPriceStep");
-      });
-    });
-  });
-  describe("parseChainData", () => {
-    describe("exact match", () => {
-      const selected = data.chains;
-      const expected = parseChainData(selected);
-      it("should match provided data", () => {
-        expect(expected).toEqual(selected);
-      });
-    });
-  });
-});
