+++ src/adapter/HttpAdapter.ts
