--- src/test/constants/tokens.ts
@@ -1,123 +1,0 @@
-export const supportedTokens = [
-  {
-    chainId: 1284,
-    address: "0xAcc15dC74880C9944775448304B263D191c6077F",
-    name: "Wrapped GLMR",
-    symbol: "WGLMR",
-    decimals: 18,
-    crosschain: false,
-    commonKey: "uwglmr",
-    logoURI:
-      "https://assets.coingecko.com/coins/images/22459/small/glmr.png?1641880985"
-  },
-  {
-    chainId: 1284,
-    address: "0xCa01a1D0993565291051daFF390892518ACfAD3A",
-    name: "Axelar USD Coin",
-    symbol: "axlUSDC",
-    decimals: 6,
-    commonKey: "uusdc",
-    crosschain: true,
-    logoURI:
-      "https://assets.coingecko.com/coins/images/6319/small/USD_Coin_icon.png?1547042389"
-  },
-  {
-    chainId: 43114,
-    address: "0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7",
-    name: "Wrapped AVAX",
-    symbol: "WAVAX",
-    decimals: 18,
-    crosschain: false,
-    commonKey: "uawax",
-    logoURI:
-      "https://raw.githubusercontent.com/pangolindex/tokens/main/assets/0xB31f66AA3C1e785363F0875A1B74E27b85FD66c7/logo_24.png"
-  },
-  {
-    chainId: 43114,
-    address: "0x49D5c2BdFfac6CE2BFdB6640F4F80f226bc10bAB",
-    name: "Wrapped ETH.e",
-    symbol: "WETH.e",
-    decimals: 18,
-    crosschain: false,
-    commonKey: "uwethe",
-    logoURI:
-      "https://assets.coingecko.com/coins/images/2518/small/weth.png?1628852295"
-  },
-  {
-    chainId: 43114,
-    address: "0xfaB550568C688d5D8A52C7d794cb93Edc26eC0eC",
-    name: "Axelar USD Coin",
-    symbol: "axlUSDC",
-    decimals: 6,
-    commonKey: "uusdc",
-    crosschain: true,
-    logoURI:
-      "https://assets.coingecko.com/coins/images/6319/small/USD_Coin_icon.png?1547042389"
-  },
-  {
-    chainId: 43114,
-    address: "0xd586E7F844cEa2F87f50152665BCbc2C279D8d70",
-    name: "DAI.e",
-    symbol: "DAI.e",
-    decimals: 18,
-    commonKey: "udaie",
-    crosschain: false,
-    logoURI:
-      "https://assets.coingecko.com/coins/images/9956/thumb/4943.png?1636636734"
-  },
-  {
-    chainId: 43114,
-    address: "0xA7D7079b0FEaD91F3e65f86E8915Cb59c1a4C664",
-    name: "USDC.e",
-    symbol: "USDC.e",
-    decimals: 6,
-    commonKey: "uusdce",
-    crosschain: false,
-    logoURI:
-      "https://assets.coingecko.com/coins/images/6319/small/USD_Coin_icon.png?1547042389"
-  },
-  {
-    chainId: 1,
-    address: "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
-    name: "Wrapped ETH",
-    symbol: "WETH",
-    decimals: 18,
-    crosschain: false,
-    commonKey: "uweth",
-    logoURI:
-      "https://assets.coingecko.com/coins/images/2518/small/weth.png?1628852295"
-  },
-  {
-    chainId: 1,
-    address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
-    name: "USD Coin",
-    symbol: "USDC",
-    decimals: 6,
-    commonKey: "uusdc",
-    crosschain: true,
-    logoURI:
-      "https://assets.coingecko.com/coins/images/6319/small/USD_Coin_icon.png?1547042389"
-  },
-  {
-    chainId: 1,
-    address: "0xdAC17F958D2ee523a2206206994597C13D831ec7",
-    name: "USD Tether",
-    symbol: "USDT",
-    decimals: 6,
-    commonKey: "uusdt",
-    crosschain: false,
-    logoURI:
-      "https://assets.coingecko.com/coins/images/325/thumb/Tether-logo.png?1598003707"
-  },
-  {
-    chainId: 1,
-    address: "0x6B175474E89094C44Da98b954EedeAC495271d0F",
-    name: "DAI",
-    symbol: "DAI",
-    decimals: 18,
-    commonKey: "udai",
-    crosschain: false,
-    logoURI:
-      "https://assets.coingecko.com/coins/images/9956/thumb/4943.png?1636636734"
-  }
-];
