--- src/v2/constants/index.ts
@@ -1,4 +1,0 @@
-export const uint256MaxValue =
-  "115792089237316195423570985008687907853269984665640564039457584007913129639935";
-
-export const nativeTokenConstant = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
